requirements.txt
# PicTrans 依赖清单

# Web框架
flask>=2.3.0
flask-cors>=4.0.0

# HTTP客户端
httpx>=0.24.0

# 图像处理
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# 工具
python-dotenv>=1.0.0

# 可选: LaMa模型支持 (高质量模式)
# torch>=2.0.0
# torchvision>=0.15.0
